+++ application/account-management/WebApp/shared/translations/locale/en-US.po
@@ -14,16 +14,16 @@ msgstr ""
 "X-Generator: @lingui/cli\n"
 
 #. js-lingui-explicit-id
+msgid "yourname@example.com"
+msgstr "yourname@example.com"
+
+#. js-lingui-explicit-id
 msgid "Account registration is not started."
 msgstr "Account registration is not started."
 
 #. js-lingui-explicit-id
 msgid "An error occured when trying to complete Account registration."
 msgstr "An error occured when trying to complete Account registration."
-
-#. js-lingui-explicit-id
-msgid "yourname@example.com"
-msgstr "yourname@example.com"
 
 #. js-lingui-explicit-id
 msgid "Email"
